Handle old system data in health check
After running sstableupgrade or sstablescrub with --migrate, the
data directory layout and filename structure will be upgraded, but
the system data will not.  In order to pass the system data
health check, the old status CF (containing the cluster name) needs
to be checked to find the saved cluster name, since the new format
system data will not be created yet.