Adds "weak" row tombstones for MV timestamp issues
The patch allow row tombstones to be either regular (as before this
patch) or weak. If a row tombstone is weak, it only applies to the row
and its content as long as the row timestamp is less than the tombstone
timestamp. As soon as the row is updated with a timestamp greater than
the one of the weak tombstone, that tombstone is removed, even if some
cells inside the row may have older timestamps.