Don't check row filter satifisfaction for 2i queries
First, the row filter conditions are alread handled elsewhere for
secondary index queries.  Second, Operator.isSatisfiedBy() is not
designed to handle all operators (IN, CONTAINS, and CONTAINS KEY are not
handled).  Right now we do not support any select queries where those
operators could be used on partition or clustering columns as part of a
row filter.  CONTAINS and CONTAINS KEY can only be used on non-frozen
collections (which can't be in the primary key) and IN restrictions are
handled outside of the row filter.